requirements.txt
kivy
reactivex
GitPython
